did some test refactoring
