+++ src/Machete.HL7.Tests/ValueDefaultExtensionsTests.cs
@@ -18,27 +18,9 @@ namespace Machete.HL7.Tests
             }))
         {
         }
-        
-        [Test]
-        public void Verify_()
-        {
-            const string message = @"MSH|^~\&|LIFTLAB||UBERMED||201701131234|||K113|P|";
-
-            Parsed<HL7Entity> parsed = Parser.Parse(message);
-
-            var query = parsed.CreateQuery(q =>
-                from x in q.Select<MSHSegment>()
-                select x);
-
-            var result = parsed.Query(query);
-
-            var messageType = result.Value.MessageType.ValueOrDefault(null);
-            
-            Assert.IsNull(messageType);
-        }
 
         [Test]
-        public void Should_()
+        public void Should_be_possible()
         {
             const string message = @"MSH|^~\&|LIFTLAB||UBERMED||201701131234||^R01|K113|P|";
 
@@ -50,9 +32,12 @@ namespace Machete.HL7.Tests
 
             var result = parsed.Query(query);
 
-            string messageType = result.Value.MessageType.Value.MessageCode.ValueOrDefault("ORU");
+            var messageType = result.Value.MessageType.Value;
+            string messageCode = messageType.MessageCode.ValueOrDefault("ORU");
             
-            Assert.AreEqual("ORU", messageType);
+            Assert.IsNotNull(messageType);
+            Assert.That(messageType.MessageCode.HasValue, Is.False);
+            Assert.AreEqual("ORU", messageCode);
         }
     }
 }
