+++ src/Machete/ValueDefaultExtensions.cs
@@ -1,8 +1,5 @@
 namespace Machete
 {
-    using System;
-
-    
     public static class ValueDefaultExtensions
     {
         /// <summary>
@@ -12,15 +9,11 @@ namespace Machete
         /// <param name="defaultValue"></param>
         /// <typeparam name="TValue"></typeparam>
         /// <returns></returns>
-        /// <exception cref="ArgumentNullException"></exception>
         public static TValue ValueOrDefault<TValue>(this Value<TValue> value)
         {
-            if (value == null || !value.HasValue)
-                return default(TValue);
-
-            return value.Value;
+            return value != null && value.HasValue ? value.Value : default(TValue);
         }
-        
+
         /// <summary>
         /// 
         /// </summary>
@@ -28,13 +21,9 @@ namespace Machete
         /// <param name="defaultValue"></param>
         /// <typeparam name="TValue"></typeparam>
         /// <returns></returns>
-        /// <exception cref="ArgumentNullException"></exception>
         public static TValue ValueOrDefault<TValue>(this Value<TValue> value, TValue defaultValue)
         {
-            if (value == null || !value.HasValue)
-                return default(TValue);
-
-            return value.HasValue ? value.Value : defaultValue;
+            return value != null && value.HasValue ? value.Value : defaultValue;
         }
     }
 }
